Move some pull request functions from models to services
